Adding VRF and VXLAN support for systemd-networkd. LP: #1764716, #1773522
