+++ .github/workflows/build-abi.yml
@@ -1,4 +1,4 @@
-name: Build
+name: Build & ABI compatibility
 
 # Controls when the action will run. Triggers the workflow on push or pull request
 # events but only for the main branch
@@ -26,8 +26,19 @@ jobs:
           sudo sed -i '/deb-src/s/^# //' /etc/apt/sources.list
           sudo apt update
           #sudo apt install lcov python3-coverage curl
+          sudo apt install software-properties-common
+          sudo add-apt-repository -y ppa:slyon/ci-netplan
+          sudo apt install abigail-tools
           sudo apt build-dep netplan.io
 
       # Runs the build
+      # TODO: move to meson build (on Jammy), once available
       - name: Run build
         run: make
+
+      # Abigail ABI checker
+      # TODO: make use of "--header-file2 src/abi.h" instead of moving abi.h
+      #       (once we can switch to a newer version of abigail-tools in CI)
+      - name: Check ABI compatibility
+        run: |
+          abidiff abi-compat/focal_0.104.xml libnetplan.so.0.0 --headers-dir2 include/ --header-file2 src/abi.h
